Sharing the code, renderColAlias and rollupexp functions
