+++ core/src/main/scala/com/yahoo/maha/core/query/hive/HiveOuterGroupByQueryGenerator.scala
@@ -37,75 +37,6 @@ abstract case class HiveOuterGroupByQueryGenerator(partitionColumnRenderer:Parti
 
     val aliasColumnMapOfRequestCols = new mutable.HashMap[String, Column]()
 
-    def renderColumnWithAlias(fact: Fact,
-                              column: Column,
-                              alias: String,
-                              requiredInnerCols: Set[String],
-                              isOuterColumn: Boolean): Unit = {
-      val name = column.alias.getOrElse(column.name)
-      val exp = column match {
-        case any if queryBuilderContext.containsColByName(name) =>
-          //do nothing, we've already processed it
-          ""
-        case DimCol(_, dt, _, _, _, _) if dt.hasStaticMapping =>
-          val renderedAlias = renderColumnAlias(alias)
-          queryBuilderContext.setFactColAliasAndExpression(alias, renderedAlias, column, Option(name))
-          s"${renderStaticMappedDimension(column)} $name"
-        case DimCol(_, dt, _, _, _, _) =>
-          val renderedAlias = renderColumnAlias(alias)
-          queryBuilderContext.setFactColAliasAndExpression(alias, renderedAlias, column, Option(name))
-          name
-        case HiveDerDimCol(_, dt, _, de, _, _, _) =>
-          val renderedAlias = renderColumnAlias(alias)
-          queryBuilderContext.setFactColAlias(alias, renderedAlias, column)
-          s"""${de.render(name, Map.empty)} $renderedAlias"""
-        case FactCol(_, dt, _, rollup, _, _, _) =>
-          dt match {
-            case DecType(_, _, Some(default), Some(min), Some(max), _) =>
-              val renderedAlias = renderColumnAlias(alias)
-              val minMaxClause = s"CASE WHEN (($name >= $min) AND ($name <= $max)) THEN $name ELSE $default END"
-              queryBuilderContext.setFactColAlias(alias, renderedAlias, column)
-              s"""${renderRollupExpression(name, rollup, Option(minMaxClause))} $renderedAlias"""
-            case IntType(_, _, Some(default), Some(min), Some(max)) =>
-              val renderedAlias = renderColumnAlias(alias)
-              val minMaxClause = s"CASE WHEN (($name >= $min) AND ($name <= $max)) THEN $name ELSE $default END"
-              queryBuilderContext.setFactColAlias(alias, renderedAlias, column)
-              s"""${renderRollupExpression(name, rollup, Option(minMaxClause))} $renderedAlias"""
-            case _ =>
-              val renderedAlias = renderColumnAlias(alias)
-              queryBuilderContext.setFactColAliasAndExpression(alias, renderedAlias, column, Option(name))
-              s"""${renderRollupExpression(name, rollup)} $name"""
-          }
-        case HiveDerDimAggregateCol(_, dt, cc, de, _, _, _) =>
-          // this col always has rollup expresion in derived expression as requirement
-          val renderedAlias = renderColumnAlias(alias)
-          queryBuilderContext.setFactColAlias(alias, renderedAlias, column)
-          s"""${renderRollupExpression(de.render(name, Map.empty), NoopRollup)} $renderedAlias"""
-        case HiveDerFactCol(_, _, dt, cc, de, annotations, rollup, _)
-          if queryContext.factBestCandidate.filterCols.contains(name) || de.expression.hasRollupExpression || requiredInnerCols(name)
-            || de.isDimensionDriven =>
-          val renderedAlias = renderColumnAlias(alias)
-          queryBuilderContext.setFactColAlias(alias, renderedAlias, column)
-          s"""${renderRollupExpression(de.render(name, Map.empty), rollup)} $renderedAlias"""
-        case HiveDerFactCol(_, _, dt, cc, de, annotations, _, _) =>
-          //means no fact operation on this column, push expression outside
-          de.sourceColumns.foreach {
-            case src if src != name =>
-              val sourceCol = fact.columnsByNameMap(src)
-              //val renderedAlias = renderColumnAlias(sourceCol.name)
-              val renderedAlias = sourceCol.alias.getOrElse(sourceCol.name)
-              renderColumnWithAlias(fact, sourceCol, renderedAlias, requiredInnerCols, isOuterColumn)
-            case _ => //do nothing if we reference ourselves
-          }
-          //val renderedAlias = renderColumnAlias(alias)
-          val renderedAlias = renderColumnAlias(alias)
-          queryBuilderContext.setFactColAliasAndExpression(alias, renderedAlias, column, Option(s"""(${de.render(renderedAlias, queryBuilderContext.getColAliasToFactColNameMap, expandDerivedExpression = false)})"""))
-          ""
-      }
-
-      queryBuilder.addFactViewColumn(exp)
-    }
-
     def generateOrderByClause(queryContext: DimFactOuterGroupByQueryQueryContext,
                               queryBuilder: QueryBuilder): Unit = {
       val model = queryContext.requestModel
@@ -193,7 +124,7 @@ abstract case class HiveOuterGroupByQueryGenerator(partitionColumnRenderer:Parti
                                 queryBuilder: QueryBuilder,
                                 queryBuilderContext: QueryBuilderContext,
                                 renderRollupExpression: (String, RollupExpression, Option[String]) => String,
-                                renderColumnWithAlias: (Fact, Column, String, Set[String], Boolean) => Unit,
+                                renderColumnWithAlias: (Fact, Column, String, Set[String], Boolean, QueryContext, QueryBuilderContext, QueryBuilder) => Unit,
                                 primitiveColsSet: mutable.LinkedHashSet[(String, Column)],
                                 noopRollupColSet: mutable.LinkedHashSet[(String, Column)]) : String = { // base: query common
 
@@ -213,7 +144,7 @@ abstract case class HiveOuterGroupByQueryGenerator(partitionColumnRenderer:Parti
         case (column, alias) =>
           val name = column.name
           val nameOrAlias = column.alias.getOrElse(name)
-          renderColumnWithAlias(fact, column, alias, Set.empty, false)
+          renderColumnWithAlias(fact, column, alias, Set.empty, false, queryContext, queryBuilderContext, queryBuilder)
           val isAggregatedDimCol = isAggregateDimCol(column)
           if (!isAggregatedDimCol) {
             if (column.isDerivedColumn) {
@@ -268,7 +199,7 @@ abstract case class HiveOuterGroupByQueryGenerator(partitionColumnRenderer:Parti
     // Render Primitive columns
     primitiveColsSet.foreach {
       case (alias:String, column: Column) =>
-        renderColumnWithAlias(fact, column, alias, Set.empty, false)
+        renderColumnWithAlias(fact, column, alias, Set.empty, false, queryContext, queryBuilderContext, queryBuilder)
     }
 
     val hasPartitioningScheme = fact.annotations.contains(HiveQueryGenerator.ANY_PARTITIONING_SCHEME)
@@ -707,23 +638,6 @@ abstract case class HiveOuterGroupByQueryGenerator(partitionColumnRenderer:Parti
           (s"'$value'", finalAlias)
         case _ => throw new UnsupportedOperationException("Unsupported Column Type")
       }
-  }
-
-  /*
-  Commonly used method
-   */
-  def renderRollupExpression(expression: String, rollupExpression: RollupExpression, renderedColExp: Option[String] = None) : String = {
-    rollupExpression match {
-      case SumRollup => s"SUM($expression)"
-      case MaxRollup => s"MAX($expression)"
-      case MinRollup => s"MIN($expression)"
-      case AverageRollup => s"AVG($expression)"
-      case HiveCustomRollup(exp) => {
-        s"(${exp.render(expression, Map.empty, renderedColExp)})"
-      }
-      case NoopRollup => s"($expression)"
-      case any => throw new UnsupportedOperationException(s"Unhandled rollup expression : $any")
-    }
   }
 
   /*
