+++ core/src/main/scala/com/yahoo/maha/core/query/hive/HiveQueryGeneratorV2.scala
@@ -104,96 +104,13 @@ class HiveQueryGeneratorV2(partitionColumnRenderer:PartitionColumnRenderer, udfS
       }
     }
 
-    def renderRollupExpression(expression: String, rollupExpression: RollupExpression, renderedColExp: Option[String] = None) : String = {
-      rollupExpression match {
-        case SumRollup => s"SUM($expression)"
-        case MaxRollup => s"MAX($expression)"
-        case MinRollup => s"MIN($expression)"
-        case AverageRollup => s"AVG($expression)"
-        case HiveCustomRollup(exp) => {
-          s"(${exp.render(expression, Map.empty, renderedColExp)})"
-        }
-        case NoopRollup => s"($expression)"
-        case any => throw new UnsupportedOperationException(s"Unhandled rollup expression : $any")
-      }
-    }
-
     def renderDerivedFactCols(derivedCols: List[(Column, String)]) = {
       val requiredInnerCols: Set[String] =
         derivedCols.view.map(_._1.asInstanceOf[DerivedColumn]).flatMap(dc => dc.derivedExpression.sourceColumns).toSet
       derivedCols.foreach {
         case (column, alias) =>
-          renderColumnWithAlias(fact: Fact, column, alias, requiredInnerCols, false)
+          renderColumnWithAlias(fact: Fact, column, alias, requiredInnerCols, false, queryContext, queryBuilderContext, queryBuilder)
       }
-    }
-
-    def renderColumnWithAlias(fact: Fact,
-                              column: Column,
-                              alias: String,
-                              requiredInnerCols: Set[String],
-                              isOuterColumn: Boolean): Unit = {
-      val name = column.alias.getOrElse(column.name)
-      val exp = column match {
-        case any if queryBuilderContext.containsColByName(name) =>
-          //do nothing, we've already processed it
-          ""
-        case DimCol(_, dt, _, _, _, _) if dt.hasStaticMapping =>
-          val renderedAlias = renderColumnAlias(alias)
-          queryBuilderContext.setFactColAliasAndExpression(alias, renderedAlias, column, Option(name))
-          s"${renderStaticMappedDimension(column)} $name"
-        case DimCol(_, dt, _, _, _, _) =>
-          val renderedAlias = renderColumnAlias(alias)
-          queryBuilderContext.setFactColAliasAndExpression(alias, renderedAlias, column, Option(name))
-          name
-        case HiveDerDimCol(_, dt, _, de, _, _, _) =>
-          val renderedAlias = renderColumnAlias(alias)
-          queryBuilderContext.setFactColAlias(alias, renderedAlias, column)
-          s"""${de.render(name, Map.empty)} $renderedAlias"""
-        case FactCol(_, dt, _, rollup, _, _, _) =>
-          dt match {
-            case DecType(_, _, Some(default), Some(min), Some(max), _) =>
-              val renderedAlias = renderColumnAlias(alias)
-              val minMaxClause = s"CASE WHEN (($name >= $min) AND ($name <= $max)) THEN $name ELSE $default END"
-              queryBuilderContext.setFactColAlias(alias, renderedAlias, column)
-              s"""${renderRollupExpression(name, rollup, Option(minMaxClause))} $renderedAlias"""
-            case IntType(_, _, Some(default), Some(min), Some(max)) =>
-              val renderedAlias = renderColumnAlias(alias)
-              val minMaxClause = s"CASE WHEN (($name >= $min) AND ($name <= $max)) THEN $name ELSE $default END"
-              queryBuilderContext.setFactColAlias(alias, renderedAlias, column)
-              s"""${renderRollupExpression(name, rollup, Option(minMaxClause))} $renderedAlias"""
-            case _ =>
-              val renderedAlias = renderColumnAlias(alias)
-              queryBuilderContext.setFactColAliasAndExpression(alias, renderedAlias, column, Option(name))
-              s"""${renderRollupExpression(name, rollup)} $name"""
-          }
-        case HiveDerDimAggregateCol(_, dt, cc, de, _, _, _) =>
-          // this col always has rollup expresion in derived expression as requirement
-          val renderedAlias = renderColumnAlias(alias)
-          queryBuilderContext.setFactColAlias(alias, renderedAlias, column)
-          s"""${renderRollupExpression(de.render(name, Map.empty), NoopRollup)} $renderedAlias"""
-        case HiveDerFactCol(_, _, dt, cc, de, annotations, rollup, _)
-          if queryContext.factBestCandidate.filterCols.contains(name) || de.expression.hasRollupExpression || requiredInnerCols(name)
-            || de.isDimensionDriven =>
-          val renderedAlias = renderColumnAlias(alias)
-          queryBuilderContext.setFactColAlias(alias, renderedAlias, column)
-          s"""${renderRollupExpression(de.render(name, Map.empty), rollup)} $renderedAlias"""
-        case HiveDerFactCol(_, _, dt, cc, de, annotations, _, _) =>
-          //means no fact operation on this column, push expression outside
-          de.sourceColumns.foreach {
-            case src if src != name =>
-              val sourceCol = fact.columnsByNameMap(src)
-              //val renderedAlias = renderColumnAlias(sourceCol.name)
-              val renderedAlias = sourceCol.alias.getOrElse(sourceCol.name)
-              renderColumnWithAlias(fact, sourceCol, renderedAlias, requiredInnerCols, isOuterColumn)
-            case _ => //do nothing if we reference ourselves
-          }
-          //val renderedAlias = renderColumnAlias(alias)
-          val renderedAlias = renderColumnAlias(alias)
-          queryBuilderContext.setFactColAliasAndExpression(alias, renderedAlias, column, Option(s"""(${de.render(renderedAlias, queryBuilderContext.getColAliasToFactColNameMap, expandDerivedExpression = false)})"""))
-          ""
-      }
-
-      queryBuilder.addFactViewColumn(exp)
     }
 
     def generateOrderByClause(queryContext: CombinedQueryContext,
@@ -233,7 +150,7 @@ class HiveQueryGeneratorV2(partitionColumnRenderer:PartitionColumnRenderer, udfS
      * Final Query
      */
 
-    val factQueryFragment = generateFactQueryFragment(queryContext, queryBuilder, renderDerivedFactCols, renderRollupExpression, renderColumnWithAlias)
+    val factQueryFragment = generateFactQueryFragment(queryContext, queryBuilderContext, queryBuilder, renderDerivedFactCols, renderRollupExpression, renderColumnWithAlias)
     generateDimSelects(dims, queryBuilderContext, queryBuilder, requestModel, fact, factViewAlias)
 
     generateOrderByClause(queryContext, queryBuilder)
